+++ saleor/payment/utils.py
@@ -176,7 +176,8 @@ def create_payment_information(
         email = cast(str, checkout.get_customer_email())
         user_id = checkout.user_id
         checkout_token = str(checkout.token)
-        checkout_metadata = checkout.metadata_storage.metadata
+        from ..checkout.utils import get_or_create_checkout_metadata
+        checkout_metadata = get_or_create_checkout_metadata(checkout)
     elif order := payment.order:
         billing = order.billing_address
         shipping = order.shipping_address